+++ kernel/server/mode/dynamic/file.go
@@ -1,8 +1,8 @@
 package dynamic
 
 import (
+	"bytes"
 	"net/http"
-	"path"
 	"strings"
 
 	"publish/client"
@@ -11,45 +11,46 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
+/* 资源文件转发 */
 func File(c *gin.Context) {
 	var (
 		request_fullpath = c.FullPath()    // 资源请求路径
 		relative_path    = c.Param("path") // 资源相对路径
-		root_path        string            // 根路径
-		file_path        string            // 文件路径
+		//  REF [golang中的io.Reader/Writer_SuPhoebe的博客-CSDN博客_golang io.reader](https://blog.csdn.net/u013007900/article/details/89126811
+		buffer    = new(bytes.Buffer) // 文件缓冲区
+		root_path string              // 根路径
 	)
 
 	/* 设置资源存放路径 */
 	switch {
 	case strings.HasPrefix(request_fullpath, config.C.Server.Static.Appearance.Path):
 		root_path = config.C.Server.Static.Appearance.Path
-		file_path = path.Join(config.C.Server.Static.Appearance.FilePath, relative_path)
 	case strings.HasPrefix(request_fullpath, config.C.Server.Static.Assets.Path):
 		root_path = config.C.Server.Static.Assets.Path
-		file_path = path.Join(config.C.Server.Static.Assets.FilePath, relative_path)
 	case strings.HasPrefix(request_fullpath, config.C.Server.Static.Emojis.Path):
 		root_path = config.C.Server.Static.Emojis.Path
-		file_path = path.Join(config.C.Server.Static.Emojis.FilePath, relative_path)
 	case strings.HasPrefix(request_fullpath, config.C.Server.Static.Export.Path):
 		root_path = config.C.Server.Static.Export.Path
-		file_path = path.Join(config.C.Server.Static.Export.FilePath, relative_path)
 	case strings.HasPrefix(request_fullpath, config.C.Server.Static.Stage.Path):
 		root_path = config.C.Server.Static.Stage.Path
-		file_path = path.Join(config.C.Server.Static.Stage.FilePath, relative_path)
 	case strings.HasPrefix(request_fullpath, config.C.Server.Static.Widgets.Path):
 		root_path = config.C.Server.Static.Widgets.Path
-		file_path = path.Join(config.C.Server.Static.Widgets.FilePath, relative_path)
 	}
 
-	// fmt.Println(request_fullpath, relative_path, file_path)
-
-	/* 下载资源文件 */
+	/* 获取资源文件 */
 	// REF [下载 - Req](https://req.cool/zh/docs/tutorial/download/)
-	if _, err := client.C.R().SetOutputFile(file_path).Get(config.C.Siyuan.Server + root_path + relative_path); err != nil {
+	if response, err := client.C.R().SetOutput(buffer).Get(config.C.Siyuan.Server + root_path + relative_path); err != nil {
 		c.Status(http.StatusNotFound)
-		return
+	} else {
+		// [从 reader 读取数据 | Gin Web Framework](https://gin-gonic.com/zh-cn/docs/examples/serving-data-from-reader/)
+		c.Data(http.StatusOK, response.GetContentType(), buffer.Bytes())
+		// c.DataFromReader(http.StatusOK, response.ContentLength, response.GetContentType(), buffer, nil)
 	}
 
 	/* 返回资源文件 */
-	c.File(file_path)
+	// REF [gin框架中多种数据格式返回请求结果 - 码农教程](http://www.manongjc.com/detail/26-epvqgdbxefbfjfd.html)
+	// if _, err := c.Writer.Write(buffer.Bytes()); err != nil {
+	// 	status.S.StatusPublishServerError(c)
+	// }
+
 }